% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefCI.R
\name{coefCI}
\alias{coefCI}
\title{Confidence Intervals for Linear Regression Coefficients}
\usage{
coefCI(mod, alpha = 0.05, approach = "asymptotic", B = 1000)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}

\item{alpha}{Number between 0 and 1 specifying the significance level
\eqn{\alpha} to obtain the \eqn{1-\alpha} confidence intervals. Default is
0.05 for a 95\% confidence interval.}

\item{approach}{A specification of which approach to use to obtain the
confidence intervals. Possible values are \code{"asymptotic"} (default) and
\code{"bootstrap"} for asymptotic and bootstrap confidence intervals,
respectively.}

\item{B}{Number of bootstrap replications to use for a bootstrap confidence
interval. Default is 1000.}
}
\value{
A matrix with columns giving lower and upper confidence limits for
each parameter. These will be labelled as \code{alpha}/2 and 1-\code{alpha}/2
in \% (by default 2.5\% and 97.5\%).
}
\description{
Obtain confidence intervals for the linear regression model
coefficients \eqn{\beta}, estimated by the function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
coefCI(mod1) # asymptotic confidence intervals
coefCI(mod1, approach = "bootstrap") # bootstrap confidence intervals

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
coefCI(mod2, alpha = 0.1, approach = "asymptotic")
coefCI(mod2, alpha = 0.1, approach = "bootstrap", B = 500)
}
