% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ftest.R
\name{Ftest}
\alias{Ftest}
\title{F-test for Linear Regression Model}
\usage{
Ftest(mod)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}
}
\value{
A \code{list} containing the following components:
\describe{
     \item{F-statistic}{F-statistic for the overall significance of the model}
     \item{df1}{Numerator degrees of freedom}
     \item{df2}{Denominator degrees of freedom}
     \item{p-value}{p-value for the F-test of overall significance}
}
}
\description{
Obtain F-test results for the overall significance of a linear
regression model fitted by the function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
Ftest(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
Ftest(mod2)
}
