% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE.R
\name{MSPE}
\alias{MSPE}
\title{Mean Square Prediction Error}
\usage{
MSPE(mod)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}
}
\value{
MSPE of the linear regression model \code{mod}.
}
\description{
Obtain Mean Square Prediction Error (MSPE) of a linear
regression model fitted by the function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
MSPE(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
MSPE(mod2)
}
