% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Estimating Linear Regression Model}
\usage{
regress(y, x, data = NULL, intercept = TRUE)
}
\arguments{
\item{y}{A numeric \code{vector} containing the sample data of the dependent
variable or, if argument \code{data} is specified, a character \code{vector}
of length 1 with the name of the dependent variable.}

\item{x}{A numeric \code{vector} or \code{matrix} containing the sample data
of the independent variables or, if argument \code{data} is specified, a
character \code{vector} with the names of the independent variables.}

\item{data}{An optional \code{data.frame} or \code{matrix} with column names,
containing the variables specified in \code{y} and \code{x}.}

\item{intercept}{A \code{logical}. If \code{TRUE} (default), a column of 1s
is added to \code{x} and the intercept coefficient is estimated along with
the slope coefficients.}
}
\value{
\code{regress} returns an object of class \code{"linreg"}.

The functions \code{\link{results}} and \code{\link{residualPlots}} are used to
obtain and print a summary and analysis of the regression results and plot the
diagnostic plots of the model residuals. The functions \code{\link{coefCI}},
\code{\link{Ftest}}, \code{\link{MSPE}}, \code{\link{Rsquared}} and
\code{\link{RsquaredAdj}} extract various useful statistics of the values
returned by \code{regress}.

An object of class \code{"linreg"} is a \code{list} containing the
following components:
\describe{
     \item{coefficients}{Estimated coefficients of the linear regression model}
     \item{error.variance}{Estimated residual variance}
     \item{beta.variance}{Estimated variance-covariance matrix of the
     estimated coefficients}
     \item{residuals}{Model residuals, that is response minus fitted values}
     \item{fitted.values}{The fitted values}
     \item{y}{The response used}
     \item{x}{The model matrix used}
     \item{n}{Number of observations used}
     \item{p}{Number of parameters estimated}
     \item{df}{The residual degrees of freedom}
}
}
\description{
Estimate the coefficient vector \eqn{\beta} for a linear
regression model \eqn{y = X'\beta + \epsilon} using ordinary least squares
and obtain the variance-covariance matrix of the estimated \eqn{\beta} and the
estimate of the residual variance \eqn{\sigma^2}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
mod1$coefficients
mod1$error.variance
mod1$beta.variance

# alternative way to specify the same model
mod2 <- regress("mpg", "wt", data = mtcars)
mod2$coefficients

mod3 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
mod3$coefficients
mod3$beta.variance

# omitting intercept
mod4 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars, intercept = FALSE)
mod4$coefficients
mod4$beta.variance
}
