% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualPlots.R
\name{residualPlots}
\alias{residualPlots}
\title{Residual Plots for Linear Regression Model}
\usage{
residualPlots(mod, which = NULL)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}

\item{which}{A number (1, 2, 3 or 4) indicating which plot is required.
If \code{NULL} (the default), the first three plots are returned in a single
figure. See the Description above for which number corresponds to which plot.}
}
\description{
Obtain diagnostic residual plots for the linear regression model
estimated by the function \code{\link{regress}}. Four plots (selectable by
\code{which}) are available: (1) a plot of residuals against fitted values,
(2) a Normal Q-Q plot of residuals, (3) a histogram of residuals and (4) a
density plot of residuals. By default, the first three are provided in a
single figure.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
residualPlots(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
residualPlots(mod2)
residualPlots(mod2, 4)
}
