% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Results of Linear Regression Model}
\usage{
results(mod, print = TRUE)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}

\item{print}{A \code{logical}. If \code{TRUE} (default), the summary of the
results is printed. If \code{FALSE}, printing is suppressed. Use
\code{print = FALSE} if you want to assign the results to an object without
printing.}
}
\value{
The function \code{results} computes and prints (if \code{print = TRUE})
a list of summary statistics of the fitted linear regression model and returns
a list containing the following components:
\describe{
     \item{results}{A matrix containing columns \code{Estimate}, \code{Std. Error},
     \code{t value} and \code{Pr(>|t|)} with the estimated model coefficients,
     their standard errors, corresponding t statistics and p-values for the
     significance of the estimated coefficients, respectively}
     \item{resid.std.error}{Estimated residual standard error}
     \item{mspe}{Mean Square Prediction Error of the estimated model}
     \item{Rsquared}{R-squared of the model}
     \item{AdjRsquared}{Adjusted R-squared of the model}
     \item{Fstatistic}{F-statistic for the overall significance of the model}
     \item{df1}{Numerator degrees of freedom for the F-test of overall
     significance}
     \item{df2}{Denominator degrees of freedom for the F-test of overall
     significance}
     \item{pvalue}{p-value for the F-test of overall significance}
}
}
\description{
Obtain summary results of a linear regression model fitted by
the function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
results(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
results(mod2)

# assign results without printing
mod2_res <- results(mod2, print = FALSE)
}
