% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{R-squared of Linear Regression Model}
\usage{
Rsquared(mod)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}
}
\value{
R-squared of the linear regression model \code{mod}.
}
\description{
Obtain R-squared of a linear regression model fitted by the
function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
Rsquared(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
Rsquared(mod2)
}
