% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsquaredAdj.R
\name{RsquaredAdj}
\alias{RsquaredAdj}
\title{Adjusted R-squared of Linear Regression Model}
\usage{
RsquaredAdj(mod)
}
\arguments{
\item{mod}{An object of class \code{"linreg"} returned by the function
\code{\link{regress}}.}
}
\value{
Adjusted R-squared of the linear regression model \code{mod}.
}
\description{
Obtain Adjusted R-squared of a linear regression model fitted
by the function \code{\link{regress}}.
}
\examples{
mod1 <- regress(mtcars$mpg, mtcars$wt)
RsquaredAdj(mod1)

mod2 <- regress("mpg", c("wt", "disp", "hp"), data = mtcars)
RsquaredAdj(mod2)
}
